+++ app/test/e2e/ct-oauth/ct-oauth-login-json.spec.js
@@ -35,7 +35,7 @@ describe('Auth endpoints tests', () => {
             .set('Content-Type', 'application/json');
 
         response.status.should.equal(200);
-        response.header['content-type'].should.equal('text/html; charset=utf-8');
+        response.should.be.html;
         response.redirects.should.be.an('array').and.length(1);
         response.should.redirectTo(/\/auth\/login$/);
     });
@@ -75,7 +75,7 @@ describe('Auth endpoints tests', () => {
             .set('Content-Type', 'application/json');
 
         response.status.should.equal(401);
-        response.header['content-type'].should.equal('application/json; charset=utf-8');
+        response.should.be.json;
         response.body.should.have.property('errors').and.be.an('array');
         response.body.errors[0].status.should.equal(401);
         response.body.errors[0].detail.should.equal('Not logged in');
@@ -87,7 +87,7 @@ describe('Auth endpoints tests', () => {
             .set('Content-Type', 'application/json');
 
         response.status.should.equal(401);
-        response.header['content-type'].should.equal('application/json; charset=utf-8');
+        response.should.be.json;
         response.body.should.have.property('errors').and.be.an('array');
         response.body.errors[0].status.should.equal(401);
         response.body.errors[0].detail.should.equal('Invalid email or password');
@@ -102,7 +102,7 @@ describe('Auth endpoints tests', () => {
             });
 
         response.status.should.equal(401);
-        response.header['content-type'].should.equal('application/json; charset=utf-8');
+        response.should.be.json;
         response.body.should.have.property('errors').and.be.an('array');
         response.body.errors[0].status.should.equal(401);
         response.body.errors[0].detail.should.equal('Invalid email or password');
@@ -118,7 +118,7 @@ describe('Auth endpoints tests', () => {
             });
 
         response.status.should.equal(401);
-        response.header['content-type'].should.equal('application/json; charset=utf-8');
+        response.should.be.json;
         response.body.should.have.property('errors').and.be.an('array');
         response.body.errors[0].status.should.equal(401);
         response.body.errors[0].detail.should.equal('Invalid email or password');
@@ -204,7 +204,7 @@ describe('Auth endpoints tests', () => {
             });
 
         response.status.should.equal(401);
-        response.header['content-type'].should.equal('application/json; charset=utf-8');
+        response.should.be.json;
         response.body.should.have.property('errors').and.be.an('array');
         response.body.errors[0].status.should.equal(401);
         response.body.errors[0].detail.should.equal('Invalid email or password');